split up utilities into the new utils folder
