+++ lambda_code/tests/utils/generic_utils.py
@@ -1,0 +1,68 @@
+"""Generic utils for tests"""
+
+import unittest
+from typing import Literal, Any
+from pydantic import ValidationError
+from jsonpath_ng.ext import parse
+
+
+def generate_field_location_for_questionnnaire_response(
+    link_id: str, field_type: Literal["code", "display", "system"]
+) -> str:
+    """Generate the field location string for questionnaire response items"""
+    return (
+        "contained[?(@.resourceType=='QuestionnaireResponse')]"
+        + f".item[?(@.linkId=='{link_id}')].answer[0].valueCoding.{field_type}"
+    )
+
+
+def generate_field_location_for_extension(
+    url: str, field_type: Literal["code", "display"]
+) -> str:
+    """Generate the field location string for extension items"""
+    return f"extension[?(@.url=='{url}')].valueCodeableConcept.coding[0].{field_type}"
+
+
+def test_valid_values_accepted(
+    test_instance: unittest.TestCase,
+    valid_json_data: dict,
+    field_location: str,
+    valid_values_to_test: list,
+):
+    """Test that valid json data is accepted by the model"""
+    for valid_item in valid_values_to_test:
+        # Update the value at the relevant field location to the valid value to be tested
+        valid_json_data = parse(field_location).update(valid_json_data, valid_item)
+        # Test that the valid data is accepted by the model
+        test_instance.assertTrue(test_instance.validator.validate(valid_json_data))
+
+
+def test_invalid_values_rejected(
+    test_instance: unittest.TestCase,
+    valid_json_data: dict,
+    field_location: str,
+    invalid_value: Any,
+    expected_error_message: str,
+    expected_error_type: Literal[
+        "type_error", "value_error", "type_error.none.not_allowed"
+    ],
+):
+    """
+    Test that invalid json data is rejected by the model, with an appropriate validation error
+
+    NOTE:
+    TypeErrors and ValueErrors are caught and converted to ValidationErrors by pydantic. When
+    this happens, the error message is suffixed with the type of error e.g. type_error or
+    value_error. This is why the test checks for the type of error in the error message.
+    """
+    # Create invalid json data by amending the value of the relevant field
+    invalid_json_data = parse(field_location).update(valid_json_data, invalid_value)
+
+    # Test that correct error message is raised
+    with test_instance.assertRaises(ValidationError) as error:
+        test_instance.validator.validate(invalid_json_data)
+
+    test_instance.assertTrue(
+        (expected_error_message + f" (type={expected_error_type})")
+        in str(error.exception)
+    )
